messenger/requirements.txt
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
boto3==1.10.28
botocore==1.13.28
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
Django==2.0.5
django-cors-headers==3.2.0
django-sslserver==0.21
django-storages==1.8
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jmespath==0.9.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.0
prompt-toolkit==2.0.10
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.2
pyparsing==2.4.2
pytest==5.2.0
python-dateutil==2.8.0
python-magic==0.4.15
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
